Add fetch options to customer cards
Passing an options hash to the cards of the customer class will initiate a new
request to re-fetch the cards by passing the options hash as query parameters.

This let you for example fetch the cards in chronological order:

    customer.cards(order: "chronological")

Or reverse_chronological order:

    customer.cards(order: "reverse_chronological")

Or any of the other filtering attributes: from, to, offset, limit.